refactor(mrml-core): move tag to render prelude
