+++ packages/mrml-core/src/mj_accordion_title/render.rs
@@ -3,9 +3,8 @@ use std::rc::Rc;
 
 use super::{MjAccordionTitle, NAME};
 use crate::helper::condition::negation_conditional_tag;
-use crate::helper::tag::Tag;
 use crate::prelude::hash::Map;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjAccordionTitleRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
