+++ packages/mrml-core/src/mj_body/render.rs
@@ -4,8 +4,7 @@ use std::rc::Rc;
 
 use super::MjBody;
 use crate::helper::size::Pixel;
-use crate::helper::tag::Tag;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjBodyRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
