+++ packages/mrml-core/src/mj_carousel_image/render.rs
@@ -3,9 +3,8 @@ use std::rc::Rc;
 
 use super::{MjCarouselImage, NAME};
 use crate::helper::size::Pixel;
-use crate::helper::tag::Tag;
 use crate::prelude::hash::Map;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjCarouselImageRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
