+++ packages/mrml-core/src/mj_column/render.rs
@@ -3,9 +3,8 @@ use std::rc::Rc;
 
 use super::{MjColumn, NAME};
 use crate::helper::size::{Pixel, Size};
-use crate::helper::tag::Tag;
 use crate::prelude::hash::Map;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjColumnRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
