+++ packages/mrml-core/src/mj_divider/render.rs
@@ -4,8 +4,7 @@ use std::rc::Rc;
 use super::{MjDivider, NAME};
 use crate::helper::condition::conditional_tag;
 use crate::helper::size::{Pixel, Size};
-use crate::helper::tag::Tag;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjDividerRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
