+++ packages/mrml-core/src/mj_image/render.rs
@@ -3,8 +3,7 @@ use std::rc::Rc;
 
 use super::{MjImage, NAME};
 use crate::helper::size::Pixel;
-use crate::helper::tag::Tag;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjImageRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
