+++ packages/mrml-core/src/mj_spacer/render.rs
@@ -3,8 +3,7 @@ use std::rc::Rc;
 
 use super::{MjSpacer, NAME};
 use crate::helper::size::Pixel;
-use crate::helper::tag::Tag;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjSpacerRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
