+++ packages/mrml-core/src/mj_table/render.rs
@@ -3,9 +3,8 @@ use std::rc::Rc;
 
 use super::{MjTable, NAME};
 use crate::helper::size::Pixel;
-use crate::helper::tag::Tag;
 use crate::mj_section::WithMjSectionBackground;
-use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable};
+use crate::prelude::render::{Error, Header, Render, RenderOptions, Renderable, Tag};
 
 struct MjTableRender<'e, 'h> {
     header: Rc<RefCell<Header<'h>>>,
